+++ config/konker-laravel-facade.php
@@ -1,5 +1,17 @@
 <?php
 
 return [
-    //
+    /*
+    |--------------------------------------------------------------------------
+    | Konker API Interface
+    |--------------------------------------------------------------------------
+    |
+    | These values are the URL and Token for the Konker API.
+    |
+    */
+
+    'konker_url' => env('KONKER_BASE_URL', 'https://konker.test/'),
+    'konker_token' => env('KONKER_ACCESS_TOKEN', 'thequickbrownfox'),
+
+
 ];