+++ changelog.md
@@ -6,3 +6,11 @@ All notable changes to `KonkerLaravelFacade` will be documented in this file.
 
 ### Added
 - Everything
+
+
+
+## Version 3.0
+
+### Added
+- Config File now used to store ENV vars for URL and Token
+- Config File published all the time